funcao-processadora/requirements.txt
google-cloud-aiplatform
google-cloud-storage
Flask
werkzeug